+++ docs/changelog.md
@@ -11,6 +11,7 @@ Version 0.13.0 (2025-06-02)
 ### Bug fixes and improvements
 
 - Add ability to use Shizuku for install sessions via `ShizukuPlugin` in `ackpine-shizuku` artifact. See documentation for details.
+- `ackpine-splits`: Don't fall back to `ZipInputStream` if reading of zipped file fails and throw instead.
 - Sample apps: don't restart installation if app was launched through an `ACTION_VIEW` intent when launching from recent apps history.
 - Sample apps: fix reappearing install session when last install session is removed.